pagination: use a set of CTEs when constructing the query with count
This changes how PostgreSQL analyses the query to essentially replace
the expensive window aggregation step with a slightly less expensive
regular aggregation step that can be performed in parallel with the
table scan. In practice, this improves performance 5-10% on my local
development environment, although PostgreSQL's query analyser thinks it
should be only about a quarter of the cost.