tests: allow arbitrary S3 hosts, and use minio for cassettes
As part of this work, I've chosen to streamline the `RECORD` environment
variable down to only allowing `yes` as an option: the
`passthrough`/`force` split added in #5040 doesn't really make sense to
me, and it feels unnecessary in practice. Either we're using the
cassettes or we're updating them.

I've also written some new documentation on integration tests. Right now
it's basically just covering the `vcr` pattern and how we use it, but
it's a starting point. There should be enough there for a contributor to
be able to regenerate the entire set of cassettes, should they wish to.

I've tested this against both a local minio (which has been used to
regenerate the cassettes) and a real S3 bucket. Both appear to work as
expected. Nevertheless, I strongly believe that we should make the
cassettes included in the repository be based on a local Minio in future
to improve the developer experience and make it easier to update these
for real.

One lingering thing I'm not so sure about are the actual cassette
updates themselves: minio evidently returns a bunch of response headers
that we didn't previously have encoded in the cassettes. I'm happy to
add them to the ignored headers list and regenerate the cassettes if we
want to get rid of the noise.