secret_scanning: make implementation more generic
This is basically to allow for an alternative secret scanning endpoint
for deps.dev, which will be similar but not identical to the endpoint we
have for GitHub.

Most of the diffstat here is actually new unit tests, since we're going
to be using the new implementation in more than one place, but note that
the integration test for the existing GitHub endpoint did not need to
change at all.